Products.jsx file
a. Products.jsx file: import Grid from material-ui/core
Going to type products.map and in the map() function we get each (product) => (we want to return something specific) which is a <Grid of type item with a key={product.id}> everytime you are looking through something with jsx using map() you want to use a {product.id} inside the Grid we add xs={12} which means full width in mobile devices for small devices it will take 6 spaces sm={6} which will mean two columns, for medium 4 md={4} for large we will do 3 lg={3}
b. In the Grid component we put <Product /> component, also you need to import the <Product/>.